:building_construction: Added products and productView controllers
